[IRIS-95] Spec update

+++ app/workers/user_requests/swap_process.rb
@@ -17,7 +17,6 @@ module UserRequests
           reject_user_request('Alternative unit is not available for swapping.')
         end
       else
-
         reject_user_request('Booking Is not available for swapping.')
       end
     end
@@ -58,11 +57,16 @@ module UserRequests
 
     # When processing fails the old project unit is restored to its previous state, booking detail is marked as swap rejected and then appropriate state, user request is rejected
     def reject_user_request(error_messages, alternate_project_unit_status=nil, new_booking_detail=nil)
-      new_booking_detail ||= current_booking_detail.related_booking_details.where(project_unit_id: alternate_project_unit.id).first
-      new_booking_detail.destroy
 
-      alternate_project_unit.make_available
-      alternate_project_unit.save
+      if alternate_project_unit.present?
+        new_booking_detail ||= current_booking_detail.related_booking_details.where(project_unit_id: alternate_project_unit.id).first
+        alternate_project_unit.make_available
+        alternate_project_unit.save
+      end
+
+      new_booking_detail.destroy if new_booking_detail
+
+
 
       user_request.reason_for_failure = error_messages
